requirements/dev.txt
Django==2.2.10
mysql-connector==2.2.9
pylint==2.4.4
python-decouple==3.1
python-dotenv==0.10.3
django-crispy-forms==1.9.0
graphene-django==2.2.0
django-filter==2.0.0
django-graphql-jwt==0.1.5
bpython
